services/log-collector/requirements.txt
Flask
Flask-CORS
pandas
pymongo
python-dotenv
